move FeedCachePolicy to its own file
