+++ PokemonEncyclopedia/PokemonEncyclopedia/FeedCache/LocalFeedLoader.swift
@@ -7,25 +7,6 @@
 
 import Foundation
 
-
-private final class FeedCachePolicy{
-	private init() {}
-	private static let calendar = Calendar(identifier: .gregorian)
-
-	private static var maxcacheAgeInDays: Int {
-		return 7
-	}
-	
-	static func validate(_ timestamp: Date, against date: Date) -> Bool{
-		guard let maxCacheAge = calendar.date(byAdding: .day, value: maxcacheAgeInDays, to: timestamp) else {
-			return false
-		}
-		
-		return date < maxCacheAge
-	}
-	
-}
-
 public final class LocalFeedLoader {
 	private let store:FeedStore
 	private let currentDate: ()-> Date
